添加分数计算
添加分数计算

+++ cocos2d-x-2.2.2/projects/motoGame/Classes/Model/GameModel.h
@@ -36,7 +36,7 @@ public:
     void appendEnergy(float energy);
     
     float getEnergyRatio();
-    float getGameScore();
+    int getGameScore();
     
 private:
     void startUseEnergy();
